Extract common hooks and actions from error controllers
