+++ app/controllers/qualifications/errors_controller.rb
@@ -1,20 +1,9 @@
 # frozen_string_literal: true
 module Qualifications
   class ErrorsController < QualificationsInterfaceController
-    skip_before_action :verify_authenticity_token
+    include HttpErrorHandling
+
     skip_before_action :authenticate_user!
     skip_before_action :handle_expired_token!
-
-    def not_found
-      render 'not_found', status: :not_found
-    end
-
-    def unprocessable_entity
-      render 'unprocessable_entity', status: :unprocessable_entity
-    end
-
-    def internal_server_error
-      render 'internal_server_error', status: :internal_server_error
-    end
   end
 end
